requirements.txt
anndata==0.12.1
h5py==3.8.0
hydra-core==1.3.2
jgraph==0.2.1
numpy==1.21.6
omegaconf==2.3.0
pandas==1.3.5
pronto==2.7.0
scanpy==1.11.4
scikit_learn==1.0.2
scipy==1.7.3
spektral==1.3.1
tensorflow_gpu==2.9.0
tensorflow_probability==0.17.0
tqdm==4.64.0
